costEstimation New

+++ app/Http/Controllers/Admin/CostEstimationController.php
@@ -5,9 +5,12 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\CostEstimationDetail;
 use App\Models\CostEstimationCalculation;
+use App\Models\MasterCalculation;
 use Illuminate\Http\Request;
-use Response;
+use Illuminate\Http\Response;
 use Yajra\DataTables\Facades\DataTables;
+
+use function PHPSTORM_META\type;
 
 class CostEstimationController extends Controller
 {
@@ -17,8 +20,29 @@ class CostEstimationController extends Controller
     {
        $data = CostEstimationDetail::select('cost_estimation_detail.*')->where('status','=','1')->orderBy('id', 'asc')->get()->toArray();
        
-        return view('admin.pages.admin-cost-estimation-single-view',compact('data'));
+       return view('admin.pages.admin-cost-estimation-single-view',compact('data'));
+        // return view('admin.pages.abc',compact('data'));
         
+        
+    }
+    public function masterCalculation(Request $request)
+    {
+        // return $request->type_id;
+        if($request->component_id && $request->type_id)
+        {
+            // return "1";
+            $total_price = 0;
+            $total_sum = 0;
+            $data = MasterCalculation::where('component_id',$request->component_id)->where('type_id',$request->type_id)->first()->toArray();
+            
+          $total_price =  (int)$data['detail_price'] + (int)$data['statistic_price'] +  (int)$data['cad_cam_price'] +  (int)$data['logistic_price'];
+          $total_sum =  (int)$data['detail_sum'] + (int)$data['statistic_sum'] +  (int)$data['cad_cam_sum'] +  (int)$data['logistic_sum'];
+            
+            // print_r(var_dump((int)$total_price));die;
+            $data['total_price'] =   $total_price;
+            $data['total_sum'] =   $total_sum;
+            return response(['status' => true, 'data' => $data ,'msg' => trans('Employee Created')], Response::HTTP_OK);
+        }
     }
     public function getEstimate(Request $request)
     {
@@ -43,13 +67,16 @@ class CostEstimationController extends Controller
         
         if($request->key)
         {
-            
+           
             $data =CostEstimationDetail::where('id',$request->key)->first();
             
             if($data)
             {
+
                 $data->contact = $request['contact'];
                 $data->date = $request['enquiry_date'];
+                $data->complexity_val = $request['complexity_val'];
+                $data->complexity_total = $request['complexity_total'];
                 $data->status = 1;
                 $data->save();
 
@@ -88,20 +115,24 @@ class CostEstimationController extends Controller
                     }
                 }
             }
-            return redirect()->route('cost-estimation-single-view');
+            // return redirect()->route('cost-estimation-single-view');
+            return response(['status' => true, 'msg' => trans('module.updated')], Response::HTTP_OK);
 
         }
         else
         {
-          
+        // print_r($request->all());die();
             $coustEstimate =new CostEstimationDetail;
                 $coustEstimate->contact = $request['contact'];
                 $coustEstimate->date = $request['enquiry_date'];
+                $coustEstimate->complexity_val = $request['complexity_val'];
+                $coustEstimate->complexity_total = $request['complexity_total'];
+                
                 $coustEstimate->status = 1;
                 $coustEstimate->save();
             if( count($request->addmore ) > 0){
                         foreach($request->addmore as  $row ) {
-                       
+                            
                           
                         $calcData = [
                             'Component' => $row['component'],
@@ -121,13 +152,13 @@ class CostEstimationController extends Controller
                             
                         ];
                         $res = $coustEstimate->CostEstimationCalculations()->create($calcData);
-                
+                        
                       
                     
                 }
                 }
-              
-                return redirect()->route('cost-estimation-single-view');
+                return response(['status' => true, 'msg' => trans('module.inserted')], Response::HTTP_OK);
+                // return redirect()->route('cost-estimation-single-view');
         }
         
        return back();
@@ -136,11 +167,28 @@ class CostEstimationController extends Controller
     {
       
         $data = CostEstimationDetail::find($id->id);
+        if (empty($data)) {
+            return response(['status' => false, 'msg' => trans('module.item_not_found')], Response::HTTP_NOT_FOUND);
+        }
         $data->status = 0;
         $data->save();
         $data->delete();
+        return response(['status' => true, 'msg' => trans('module.deleted')], Response::HTTP_OK);
 
     }
+    // public function delete_role($id)
+    // {
+        
+    //      $module = Role::find($id);
+    //     if (empty($module)) {
+    //         return response(['status' => false, 'msg' => trans('module.item_not_found')], Response::HTTP_NOT_FOUND);
+    //     }
+    //     $module->status = 2;
+    //     $module->save();
+    //     $module->delete();
+    //     return response(['status' => true, 'msg' => trans('module.deleted')], Response::HTTP_OK);
+    // }
+  
 
     public function costEstimationEdit(Request $id)
     {
@@ -148,6 +196,8 @@ class CostEstimationController extends Controller
         $arr['detail'] = CostEstimationDetail::where('id',$id->id)->first()->toArray();
         $arr['calculation'] = CostEstimationCalculation::where('estimation_detail_id',$id->id)->get()->toArray();
        
+
+        // $arr['detail']
         $detail_price = 0;
         $detail_sum = 0;
 
@@ -196,7 +246,9 @@ class CostEstimationController extends Controller
 
         $arr['total_price'] = $total_price;
         $arr['total_sum'] = $total_sum;
+
      
+        // $total_price/
      
         return response()->json(['data' => $arr]);
